+++ Sources/VComponents/Components/Modals (Containers)/Modal/VModal.swift
@@ -129,7 +129,7 @@ struct VModal<Content>: View
     // MARK: Lifecycle Animations
     private func animateIn() {
         withAnimation(
-            appearance.appearAnimation?.toSwiftUIAnimation,
+            appearance.appearAnimation,
             { isPresentedInternally = true }
         )
     }
@@ -138,7 +138,7 @@ struct VModal<Content>: View
         completion: @escaping () -> Void
     ) {
         withAnimation(
-            appearance.disappearAnimation?.toSwiftUIAnimation,
+            appearance.disappearAnimation,
             { isPresentedInternally = false },
             completion: completion
         )