+++ Sources/VComponents/Components/Modals (Containers)/SideBar/VSideBar.swift
@@ -132,7 +132,7 @@ struct VSideBar<Content>: View where Content: View {
     // MARK: Lifecycle Animations
     private func animateIn() {
         withAnimation(
-            appearance.appearAnimation?.toSwiftUIAnimation,
+            appearance.appearAnimation,
             { isPresentedInternally = true }
         )
     }
@@ -140,7 +140,7 @@ struct VSideBar<Content>: View where Content: View {
     private func animateOut(
         completion: @escaping () -> Void
     ) {
-        let animation: BasicAnimation? = {
+        let animation: Animation? = {
             if isBeingDismissedFromSwipe {
                 appearance.swipeDismissAnimation
             } else {
@@ -149,7 +149,7 @@ struct VSideBar<Content>: View where Content: View {
         }()
 
         withAnimation(
-            animation?.toSwiftUIAnimation,
+            animation,
             { isPresentedInternally = false },
             completion: completion
         )