+++ Sources/VComponents/Components/Modals (Notifications)/Toast/VToast.swift
@@ -202,7 +202,7 @@ struct VToast: View {
         playHapticEffect()
 
         withAnimation(
-            appearance.appearAnimation?.toSwiftUIAnimation,
+            appearance.appearAnimation,
             { isPresentedInternally = true },
             completion: dismissFromTimeout
         )
@@ -211,7 +211,7 @@ struct VToast: View {
     private func animateOut(
         completion: @escaping () -> Void
     ) {
-        let animation: BasicAnimation? = {
+        let animation: Animation? = {
             if isBeingDismissedFromSwipe {
                 appearance.swipeDismissAnimation
             } else {
@@ -220,7 +220,7 @@ struct VToast: View {
         }()
 
         withAnimation(
-            animation?.toSwiftUIAnimation,
+            animation,
             { isPresentedInternally = false },
             completion: completion
         )